Feat: 갤러리 라우팅 작업
