+++ src/components/gallery/Business.tsx
@@ -1,0 +1,5 @@
+const Business = () => {
+  return <div>협력사 내용</div>;
+};
+
+export default Business;
