+++ src/components/gallery/JobPosting.tsx
@@ -1,0 +1,5 @@
+const JobPosting = () => {
+  return <div>채용공고 내용</div>;
+};
+
+export default JobPosting;
