+++ src/components/gallery/OfficePhoto.tsx
@@ -1,0 +1,5 @@
+const OfficePhoto = () => {
+  return <div>내부 사진 내용</div>;
+};
+
+export default OfficePhoto;
